collapse filters.predicate and filters.programmable into filters.python. Push all tests into a single file
