+++ test/data/plang/from-module.json.in
@@ -2,7 +2,7 @@
   "pipeline":[
     "@CMAKE_SOURCE_DIR@/test/data/las/1.2-with-color.las",
     {
-      "type": "filters.predicate",
+      "type": "filters.python",
       "script": "@CMAKE_SOURCE_DIR@/test/data/plang/test1.py",
       "module": "anything",
       "function": "fff"
